Added graphics drawing and repaired image control bug.

+++ sem1/BrezenkhemAlgorithm.h
@@ -10,15 +10,15 @@ ref class BrezenkhemAlgorithm
 {
 protected:
 	//cli::array<Tuple<float, float>^>^ points; 
-	System::Collections::Generic::List<Tuple<float, float>^>^ points;
+	System::Collections::Generic::List<Tuple<int, int>^>^ points;
 
 public:
 	BrezenkhemAlgorithm();
 	~BrezenkhemAlgorithm();
-	void PushPoint(float x, float y);
-	void PutPointsToEllipseSet(float x, float y, float x0, float y0);
-	void BrezenkhemLine(float x0, float y0, float x1, float y1);
-	void BrezenkhemCircle(float x0, float y0, float radius);
-	void BrezenkhemEllipse(float x0, float y0, float x_radius, float y_radius);
+	void PushPoint(int x, int y);
+	void PutPointsToEllipseSet(int x, int y, int x0, int y0);
+	void BrezenkhemLine(int x0, int y0, int x1, int y1);
+	void BrezenkhemCircle(int x0, int y0, int radius);
+	void BrezenkhemEllipse(int x0, int y0, int x_radius, int y_radius);
 };
