+++ sem1/BrezenkhemDrawing.h
@@ -8,7 +8,7 @@ ref class BrezenkhemDrawing :
 	public BrezenkhemAlgorithm
 {
 private:
-
+	
 	void Draw();
 
 	System::Windows::Forms::PictureBox^  pictureBox;
@@ -18,14 +18,12 @@ public:
 	~BrezenkhemDrawing();
 	System::Windows::Forms::PictureBox^ getPictureBox();
 
-	void DrawLine(float x0, float y0, float x1, float y1);
+	void DrawLine(int x0, int y0, int x1, int y1);
 	void DrawLine(Point^ point0, Point^ point1);
 
-	void DrawCircle(float x0, float y0, float radius);
-	void DrawCircle(Point^ center, float radius);
+	void DrawCircle(int x0, int y0, int radius);
+	void DrawCircle(Point^ center, int radius);
 
-	void DrawEllipse(float x0, float y0, float width, float height);
-	void DrawEllipse(Point^ center, float width, float height);
+	void DrawEllipse(int x0, int y0, int width, int height);
+	void DrawEllipse(Point^ center, int width, int height);
 };
-
-
